Updated to remove some memory leaks(very minor ones, but they were leaks) Optimized the auto miner to use loops. Allowed realtime customization of the autominer. Made saving map files and logging optional, you can use it as a pure proxy now.

+++ PacketLogin.h
@@ -42,7 +42,7 @@ public:
 
 	void Print(FILE *fp)
 	{
-		fprintf(fp, "Login( version = %d, name = %s, password = %s, mapSeed = %lld, dimension = %d )", version, name, password, mapSeed, (int)dimension);
+		fprintf(fp, "Login( version = %d, name = %s, password = %s, mapSeed = %lld, dimension = %d )", version, name, password, (long long int)mapSeed, (int)dimension);
 	}
 
 	void Process(bool to_server)
